fixing search item layout

+++ app/src/main/java/com/groundzero/github/feature/search/ui/SearchViewModel.kt
@@ -15,13 +15,14 @@ class SearchViewModel @Inject constructor(private val repository: SearchReposito
     private val pageLiveData = MutableLiveData<Int>()
     private var page = 1
     val repoResult = Transformations.switchMap(queryLiveData) { query ->
-        println(page)
         repository.searchQuery(query, page, 20, SearchSort.FORKS)
     }
 
-    fun searchRepo(queryString: String) {
-        repository.deleteData()
-        queryLiveData.postValue(queryString)
+    fun searchRepo(queryString: String?) {
+        if (queryString != null) {
+            repository.deleteData()
+            queryLiveData.postValue(queryString)
+        }
     }
 
     fun nextPage() {
